Adicionando comentários para resolver alguns problemas no código

+++ ControleDeEstoqueApi/Domain/Models/Fabricante.cs
@@ -6,8 +6,10 @@ namespace ControleDeEstoqueApi.Domain.Models
     [Table("fabricante")]
     public class Fabricante
     {
+        // TODO: Criar um Id para cada classe como Primary Key.
+
         [Key]
-        internal long cod_fab { get; set; }
+        internal int cod_fab { get; set; }
 
         [MaxLength(255)]
         internal string nome_fab { get; set; }
